Update event handling functionality
